[Xamarin.Android.Build.Tasks] DesignTimeBuild Improvements  (#1437)
We had a report from @garuma that the `Resource.designer.cs` was not
being updated correctly if the NuGet packages were not restored when
opening the project. This is a user preference option apparently.
This is due to when the Design Time Build (DTB) is run. The main DTB
is run on project load. The `ManagedResourceParser` would then run on
the available resources, which would NOT include those from the NuGet
packages as they were not restored.

When the NuGet packages are finally restored the designer is not
updated because none of the inputs changes. On a restore NuGet does
not modify the `.csproj`, or change any of the project build
properties. As a result some of our targets a skipped. 

This commit adds the various `.cache` files we use to the inputs of
`_UpdateAndroidResgen` and other targets to ensure that they get
updated when these files change. It also adds the NuGet
`$(ProjectAssetsFile)` to the list of inputs so if that file changes
we rebuild the designer on the next DTB. 

This work uncovered a problem with the way we store the `.cache`
files `$(_AndroidResourcePathsCache)`,
`$(_AndroidLibraryProjectImportsCache)` and
`$(_AndroidLibraryImportsCache)`. It turns out that they were being
saved every time the task ran, even if the contents of the files did
not change. This then caused the `_UpdateAndroidResgen` target to run
even if it did not need to.

Add an `XDocument.SaveIfChanged()` extension method which uses the
same temp file trick we use in other places to ensure we only update
the file if it changes.