[Xamarin.Android.Build.Tasks] Rework how we resolve netstandard Nugets (#1459)
Commit f7c942dd added support for replacing Reference
assemblies with runtime assemblies. However a member of
the Nuget team correctly pointed out that the Nuget does
not akways have the same name as the package its in...

So lets rework this code to do something else. We have the
full path to the reference assembly. We also have the Nuget
Project Model. This contains a list of the Paths for the
cache's being used. So we use that information to attempt to
figure out which package this reference came from. We can the
use the Nuget ProjectModel to look up the replacement runtime
assembly.

Again this should just skip over things if we can't find what
we want and issue the normal warning.