[Xamarin.Android.Build.Tasks] `fakeLogOpen` is ignorable
Fixes: https://devdiv.visualstudio.com/DevDiv/_workitems/edit/585941

Commit 42a4b9be special-cased the `fakeLogOpen` messages to produce
warnings, instead of their earlier behavior of producing *errors*.

This was an improvement, in that builds were now able to finish, but
now developers are being "spammed" with lots of warnings about
`fakeLogOpen`, which cannot be easily ignored.

Further special-case the `fakeLogOpen` messages so that instead of
producing warnings, we instead produce "normal" messages.