[api-compatibility] Only check stable API levels (#1458)
[xamarin-android build #915][0] reports various api-compatibility
failures, such as:

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/915/consoleText

	<h1>### API BREAK BETWEEN v8.1 and v8.1.99</h1>
	<!-- start namespace Android.Telephony --> <div>
	<h2>Namespace Android.Telephony</h2>
	<!-- start type CellIdentityCdma --> <div>
	<h3>Type Changed: Android.Telephony.CellIdentityCdma</h3>
	Modified base type: <span class='removed removed-inline removed-breaking-inline'>Java.Lang.Object</span> <span class='added '>Android.Telephony.CellIdentity</span>

We want to ignore these for now, as these changes are against an
*unstable* API level, and we don't want to have to ensure
intra-API-stability for unstable API levels. (In part becuase we
haven't enumified unstable API levels, because they're in flux.)

Update `$(STABLE_API_LEVELS)` so that it includes API-27 -- which
should have been done in f41afb61 -- and update
`make run-api-compatibility-tests` so that it only checks
`$(STABLE_FRAMEWORKS)`, not *all* frameworks.