[Xamarin.Android.Build.Tasks] Fix an issue where the 'ref' nuget is not found on windows.
The previous commit 029a8b49 fixed up some issues with resolving
the correct nuget. However on windows we get some weird pathing
issues because the project.assets.json file is using `/` as
path seperators. This is probably to be platform independent.
But our string comparision was using what ever path we had, on
windows that was `\`. So we never found the Library were we
looking for. We also need to normalize the resulting full path
to match the system path separator.