[Mono.Android] remove LINQ usage in ResourceIdManager
There were quite a few LINQ expressions in here, that could be rewritten
to simple `foreach` loops. The new code should have fewer allocations.

I also updated the usage of `MethodInfo.Invoke` to use `CreateDelegate`
instead, as it is an order of magnitude faster to call the method in
this way.

Lastly, I added a null check to the return value of `GetMethod`, because
it is *technically* possible for a class to exist with the
`ResourceDesignerAttribute` but be missing an `UpdateIdValues` method.

These changes should give us a *very slight* performance improvement to
startup.