[Xamarin.Android.Build.Tasks]  error "resource directory obj\debug\res does not exist"
Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/588716

Also relates to errors such as

	"obj\Debug-Automatic\android\bin\packaged_resources" does not exist.
	error "resource directory obj\debug\res does not exist"

We have a serious problem with our Async Task system. The fact that
the background thread is executing in a certain directory cannot be
relied upon. So we need to provide full paths to the various file
items we are using in `Aapt` as well as ensure that the Working
Directory is set for the tools we are launching on the background
threads.

In this PR we are adding a `WorkingDirectory` property to `AsyncTask`.
This will allow us to always know where the task was started. We
can then use this information to build a full path for those files
which are not rooted.

Also on advice from the msbuild team the use of `Yield` and `Reaquire`
has been reworked. Rather than yielding the MSBuild Node just before
we enter the message loop, we should yield BEFORE we start the
threads. This should in theory help reduce the directory changes, but
it did not eliminate them completely. We a result we had to also
make the changes regarding using full paths.