[mono-sdks] Fix Linux build
LLVM would fail to build on Linux because it would try to compile with a --host finishing with "apple-darwin10". The fix consist in testing if we are currently building on macOS (apple-darwin10) or Linux (linux-gnu).

We also make sure to pass to the Mono SDKs the path to `external/llvm`, thus avoiding cloning it again in `external/mono/sdks/builds/toolchains/llvm`.