[plots] Get logcat output for timing in the tasks
Lately the BCL tests started to produce a lot of logcat output, way
over the 4M limit of the buffer.

Thus we need to get the logcat output after every apk test run.

I have considered to batch the `RunTestApks` target. That would not
work very well though, as we are already batching the tasks in this
target with two related different item groups. So I moved the
collection of the logcat output to the `RunInstrumentationTests` and
`RunUITests` tasks.