build fix: do not generate API XML definitions every time.
https://github.com/xamarin/xamarin-android/pull/1491 broke build caching.
Now api-*.xml.in is generated EVERY TIME, which adds tremendous amount of
build time. And this build happens every time anything that depends on
api-*.xml.in and all its subsequent dependencies.
For example, Xamarin.Android.Build.Tasks build is suffered from it.

The cause of the problem is this:

	Inputs="@(_ApiParameterDescription)"

which is expanded as follows:

	/sources/xamarin-android/build-tools/api-xml-adjuster/../../src/Mono.Android/Profiles/api-4.params.txt;/sources/xamarin-android/build-tools/api-xml-adjuster/../../src/Mono.Android/Profiles/api-5.params.txt; ...

Note that many of those (e.g. API Level 4-9) do not exist.
For non-existent inputs, MSBuild targets always execute.

To fix the issue, add another build task that returns only existent files
and use it to determine the actual targets.

This is sadly a compromised solution that updates on ANY API level will
trigger generation of ALL API levels (but must be applied IMMEDIATELY
without excuse for "better solution". It's a BLOCKER for development.
No one can work on MSBuild tasks if this bug remains).