[windows] Fail the build if bundle download fails (#1528)
One of my biggest nightmares when developing for `xamarin-android` on
Windows is noticing when I inadvertently modified the hash used for
the bundle, or check out a newer commit where the bundle is not
available yet. In cases where the bundle gets a 404, this is merely a
warning and the build continues on. This is a problem on Windows,
since it completely depends on the bundle download.

Depending on the state of your build tree, one of the following could
happen on Windows due to a failed bundle download:

  - An `<Exec/>` task fails, due to it trying to run `make`
  - Thousands of build errors on `Mono.Android.csproj` complaining
    about missing `mscorlib` types.

Explicitly failing the build on Windows makes it clearer what went
wrong. It also fails the build faster, freeing up development/build
agent time.

I used the `$(MSBuildLastTaskResult)` built-in property to check if the
`<DownloadUri/>` task failed, even though it has
`ContinueOnError=True`. This is nice, because you get the 404 error
message as well as the newly added message.