[Xamarin.Android.Build.Tasks] Speed up ConvertResourcesCases (#1522)
We currently call the `<ConvertResourcesCases/>` task in [three][1]
[different][2] [places][3]. For the most part they process the same
files, over and over and over again. On a large project this is
totally a waste of time. If the file has been processed once and not
updated then we should not need to process it again.

[1]: https://github.com/xamarin/xamarin-android/blob/5d64fb4bbad606eb4a3478085db0ca98616f9966/src/Xamarin.Android.Build.Tasks/Xamarin.Android.Common.targets#L1951
[2]: https://github.com/xamarin/xamarin-android/blob/5d64fb4bbad606eb4a3478085db0ca98616f9966/src/Xamarin.Android.Build.Tasks/Xamarin.Android.Common.targets#L1436
[3]: https://github.com/xamarin/xamarin-android/blob/5d64fb4bbad606eb4a3478085db0ca98616f9966/src/Xamarin.Android.Build.Tasks/Xamarin.Android.Common.targets#L2011

The only exception here is `<GenerateJavaStubs/>`, which needs to be
done *after* we have compiled the app, but we *also* need to replace
any instances of `Foo.MyClass` with `<md5hash>.MyClass` which is
generated in the Java code. We thus still need to process the
resource files for the app after the main assembly has been built.

That said if we know the `UpdateAndroidResgen` target has run, then
we *don't* need to run this again in the `_CreateBaseApk` target.

This is an attempt to see if we can only process the files which have
changed between each of those calls. Alternatively we could rework
the targets so that we only have one call at the right place in the
build system (probably as part of the `UpdateAndroidResources`).

If this does not work then the next task will be to rework
`AndroidResource.UpdateXmlResources()` to *not* use `XDocument` but
try to use `XPathDocument` or `XmlReader/Writer` instead since it is
a streaming version and should be a lot quicker.

~~ Timing Results ~~

On a basic Android App which references the support libraries:


	Time Elapsed 00:00:23.40 -- without the change.
	Time Elapsed 00:00:19.43 -- with it. 

The basic app saves ~4 seconds with this change.

If we build "from clean", touch a Resource file, and build again:

**Previously**:

	5230 ms  ConvertResourcesCases                      3 calls -- From Clean
	3947 ms  ConvertResourcesCases                      2 calls -- After touch 1 file in Resources

**Now**:

	3283 ms  ConvertResourcesCases                      2 calls -- From Clean
	 119 ms  ConvertResourcesCases                      1 calls -- After touch 1 file in Resources 

The "build after touching a Resource" case is ~33x *faster*!