[generator] initial unit tests for Xml* and Managed* classes (#284)
As I begin refactoring `generator`, it looks like I will be making
changes to `ManagedClassGen` and `XmlClassGen` as well as the various
other types that use either Mono.Cecil or `XDocument` to represent a
type in `generator`.

These are the first examples of unit tests that can be written against
these classes, and are just simple smoke tests for now.

ManagedTests:

  - C# types defined in a namespace at the top of the file
  - Copies current test assembly to `%TEMP%` & load via Mono.Cecil
    - Mono.Cecil holds a lock on the file, so it was causing some
      issues with my IDE if I opened the current assembly with Cecil
  - Had to change the build action for `SupportFiles\RegisterAttribute` so
    I could use the type within the test assembly. It is still copied to the
    output directory for the integration tests.
  - Implemented a missing `ctor` for `RegisterAttribute` that was throwing
    `NotImplementedException`

XmlTests:

  - Defined XML at the top of the file
  - Use `System.Xml.Linq` to load specific elements as needed

Various smoke tests for:

  - Classes
  - Methods
  - Parameters
  - Fields
  - Interfaces

In the future, I may need to expand on these tests to make sure I
don't break things. They may also get refactored/removed one day if
we get the classes under test refactored to POCOs.