[build-tools] Allow various bundle download and unzip properties to be overridden (#1552)
The intent here is to make it easier to fetch bcl-test artifacts for
PR builds. QA's current bcl-test harness relies on the
`_DownloadAndExtractBundle` target to download and extract the
bcl-test assemblies to the correct location. However, this falls
apart when attempting to test a PR build as the mono bundle will not
be uploaded. One solution to this problem is to allow us to override
the file that is downloaded, as PR builds do upload a zip file for
the full build which also contains the relevant test assemblies.

Once the bundle download URI properties can be overridden, we quickly
run in to max path length limitations when attempting to
`<SystemUnzip/>` the entirety of the XA build on Windows. A new
property for a temporary bundle extraction location has been added
to address this.

Add the following new build-time properties:

  * `$(XABundleDownloadPrefix)`: URI, excluding filename, for where
    to download the bundle from.
  * `$(XABundleFileName)`: Filename for the bundle to download.
  * `$(XABundleUnzipPath)`: Where to extract the downloaded bundle.

Example property values:

	/p:XABundleDownloadPrefix="https://xamjenkinsartifact.blob.core.windows.net/xamarin-android/2889/xamarin-android"
	/p:XABundleFileName="xamarin.android-oss_v8.3.99.161_Darwin-x86_64_HEAD_b88ba02.zip"
	/p:XABundleUnzipPath="C:\x"