[Xamarin.Android.Build.Tasks] Deploying to Android fails with zipalign error
Fixes #1519

The change to introduce `apksigner` had a slight flaw in its
signing logic. It only ever deleted `$(ApkFileSigned)` when
rebuilding the apk. So in the case where we are splitting the
apk by abi.. we didn't delete the other abi specific apks..

This commit fixes up the targets so that we correctly remove
the abi specific apks if they are being used when `$(AndroidUseApkSigner)`
is `true`.