Generate source file with XA build paths
The generated file contains paths to the top source directory, `xabuild`,
`xabuild.exe`, current `bin` directory, current test output directory and a
constant with the current configuration name. The paths can be useful in test
projects to find various bits and pieces which are currently referred to using
paths relative to a specific location and thus not really "portable" between
various test projects.