[android-toolchain] support Windows again
Context fef4541

Since the `mono-sdks` took over the installation of `android-toolchain`,
these dependencies are now installed by `make` scripts. This
unfortunately, will not work on Windows since we allow builds with
a vanilla install of Visual Studio 2017 and a simple:
`msbuild Xamarin.Android.sln /t:Prepare` then `/t:Build`.

To fix this, I brought back the old MSBuild code in a
`android-toolchain-windows.targets` file to be used only on Windows. I
had to rearrange how things are imported a bit to get this to work
properly.

This should give us what we want:
- The `mono-sdks` install the Android toolchain on macOS/linux
- Windows still functions as it did before