[jnimarshalmethod-gen] Update the originating assembly (#300)
Update the source assembly instead of creating a *new* assembly named
`<AssemblyBaseName>-new.dll`

I expected I would need to load the originating assembly in the
separate AppDomain, but instead it was enough to load it with Cecil
with reading parameters having `ReadWrite = true`.

Because the referenced assemblies might not be writable, we need to
keep loading them as read only, thus with default `ReadWrite = false`.

For that I added new method to our Cecil assembly resolver, to add
existing assembly definition to its cache. We first read the
originating assemblies as `ReadWrite` and add them to the resolver
cache.