[Xamarin.Android.Build.Tests] Use `MSBUILD` environment variable rather than `USE_MSBUILD`.
In our old unit test system we called into msbuild/xbuild
directly. With our new system we make use of `xabuild`. This
is a wrapper around msbuild/xbuild which sets up the
xamarin-android environment.

For the old tests we passed `USE_MSBUILD` to control if we neeed
to use msbuild or xbuild. However with `xabuild` we pass the
`MSBUILD` environment varioable. This can be `msbuild` or `xbuild`
rather than a `1` or `0`.

The problem is the old variable was still being used so there
was a mismatch between what the tests were expecting and what
was being used. So take

	MSBUILD=msbuild
	USE_MSBUILD=0

The tests would expect to be using xbuild and will therefore check
for xbuild type messages and behaviour. But xabuild would be using
msbuild. This results in weird errors.

So lets get rid of the `USE_MSBUILD` usage and just use the `MSBUILD`
envionment variable to decide which system we are using.