Implement code-behind to access layout elements without FindViewById
This patch implements code-behind generation which allows one to find and access
layout elements (widgets) without having to call the `FindViewById` or
`FindFragment` APIs, instead merely accessing a C# property which will do all
the work of finding and returning the right widget.

As opposed to the old implementation, two approaches are available.

First approach
==============

Generates "binding" code, that is a stand-alone class with
properties to access the widgets. All the generated classes are placed in the
`Binding` namespace and their names follow the base name of the layout being
bound. Therefore, a class for layout named `Main.axml` will be named
`Binding.Main` and a class for layout `my_other_layout.xml` will be named
`Binding.my_other_layout`. The main advantage of this approach is that no
modifications layout files are necessary.

The class can be instantiated in a number of ways:

  * By passing instance of an activity *after* setting the activity layout in
    the usual way:

        var items = new Binding.Main (this)

  * By passing instance of a *View* to the constructor after loading the correct
    layout into the view:

        var items = new Binding.my_other_layout (some_view);

  * By changing the Activity layout setting code to call a new overload of the
    SetContentView method:

        var items = SetContentView<Binding.Main> ();

After the binding is instantiated, one can find the widgets by simply accessing
the correct property:

        Button btn = items.MyButton;

Given the following layout:

    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android">
       <Button android:id="@+id/myButton" />
       <fragment  xamarin:managedType="CommonSampleLibrary.LogFragment"
           android:name="commonsamplelibrary.LogFragment"
           android:id="@+id/log_fragment" />
       <fragment android:name="CommonSampleLibrary.LogFragment"
           android:id="@+id/secondary_log_fragment" />
    </LinearLayout>

Similar code will be generated (some elements are removed for brevity -
whitespace, line pragmas, comments and using statements):

    namespace Binding
    {
      sealed class Main : global::Xamarin.Android.Design.LayoutBinding
      {
        public override int ResourceLayoutID => global::Xamarin.Android.Tests.CodeBehindBuildTests.Resource.Layout.Main;

        public Main (global::Xamarin.Android.Design.ILayoutBindingClient client) : base (client) {}

        Button __myButton;
        public Button myButton => FindView (global::Xamarin.Android.Tests.CodeBehindBuildTests.Resource.Id.myButton, ref __myButton);

        CommonSampleLibrary.LogFragment __log_fragment;
        public CommonSampleLibrary.LogFragment log_fragment => FindFragment (global::Xamarin.Android.Tests.CodeBehindBuildTests.Resource.Id.log_fragment, ref __log_fragment);

        global::Android.App.Fragment __secondary_log_fragment;
        public global::Android.App.Fragment secondary_log_fragment => FindFragment (global::Xamarin.Android.Tests.CodeBehindBuildTests.Resource.Id.secondary_log_fragment, ref __secondary_log_fragment);

        CommonSampleLibrary.LogFragment __tertiary_log_fragment;
        public CommonSampleLibrary.LogFragment tertiary_log_fragment => FindFragment (global::Xamarin.Android.Tests.CodeBehindBuildTests.Resource.Id.tertiary_log_fragment, ref __tertiary_log_fragment);
      }
    }

`Xamarin.Android.Design.LayoutBinding` is a new class added to make it possible
to instantiate the binding for both an Activity and a View.

Second approach
===============

Builds on the first one and requires slightly more changes
to the code. It is similar to the old approach in that it generates a partial
activity class which defines a number of properties, right in the Activity, to
access the layout widgets. In order for this to work, it is first necessary to
modify the associated layout by adding two attributes to the root element of the
layout:

  * XML namespace declaration:

       xmlns:xamarin="http://schemas.xamarin.com/android/xamarin/tools"

  * Specification of *full* type names for activities which will use the
    generated code (a semicolon-separated list, at least one type is required):

       xamarin:classes="Xamarin.Android.Tests.CodeBehindBuildTests.MainActivity;Xamarin.Android.Tests.CodeBehindBuildTests.AnotherMainActivity"

After the above is done, all the activities mentioned in the `xamarin:classes`
attribute have to be modified by adding the `partial` modifier to the Activity
class declaration. No other changes are necessary as the code-behind overrides
all the `SetContentView` methods to instantiate the correct `Binding` class and
use it to seamlessly access the layout widgets. Widgets are accessed via
standard properties, e.g.:

    Button btn = MyButton;

Given the same layout as in `First approach`, with the two XML attributes added:

    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
                  xmlns:xamarin="http://schemas.xamarin.com/android/xamarin/tools"
		  xamarin:classes="Xamarin.Android.Tests.CodeBehindBuildTests.MainActivity">
       <Button android:id="@+id/myButton" />
       <fragment  xamarin:managedType="CommonSampleLibrary.LogFragment"
           android:name="commonsamplelibrary.LogFragment"
           android:id="@+id/log_fragment" />
       <fragment android:name="CommonSampleLibrary.LogFragment"
           android:id="@+id/secondary_log_fragment" />
    </LinearLayout>

We get the following code generated (the `Binding.Main` class remains exactly
the same and can still be used independently):

    namespace Xamarin.Android.Tests.CodeBehindBuildTests
    {
      partial class MainActivity
      {
        Binding.Main __layout_binding;

        public override void SetContentView (global::Android.Views.View view) {
          __layout_binding = new global::Binding.Main (view);
          bool callBase = true;
          OnSetContentView (view, ref callBase);
          if (callBase)
            base.SetContentView (view);
        }

        public override void SetContentView (global::Android.Views.View view, global::Android.Views.ViewGroup.LayoutParams @params) {
          // Code similar to above
        }

        public override void SetContentView (int layoutResID) {
          // Code similar to above
        }

        partial void OnSetContentView (global::Android.Views.View view, ref bool callBaseAfterReturn);
        partial void OnSetContentView (global::Android.Views.View view, global::Android.Views.ViewGroup.LayoutParams @params, ref bool callBaseAfterReturn);
        partial void OnSetContentView (int layoutResID, ref bool callBaseAfterReturn);

        public Button myButton => __layout_binding?.myButton;
        public CommonSampleLibrary.LogFragment log_fragment => __layout_binding?.log_fragment;
        public global::Android.App.Fragment secondary_log_fragment => __layout_binding?.secondary_log_fragment;
        public CommonSampleLibrary.LogFragment tertiary_log_fragment => __layout_binding?.tertiary_log_fragment;
      }
    }

The `OnSetContentView` partial methods may be implemented in your code to
customize `SetContentView` behavior since it's no longer possible to override
them in your "main" activity code.