[build] Generate XABuildPaths.cs for all $(CONFIGURATIONS) (#1586)
Commit e937f9b4 broke `make prepare` for the Release configuration.
Jenkins assumes that `make prepare` is a single invocation for *both*
Debug & Release builds; see e.g. `make prepare-external`:

	$(foreach conf, $(CONFIGURATIONS), ...)

Commit e937f9b4 didn't abide by this requirement, and thus only
generated `XABuildPaths.cs` for the Debug configuration. As such, when
attempting to build in the Release configuration, [things broke][0]:

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/958/

	CSC: error CS2001: Source file '…/xamarin-android/src/Xamarin.Android.Build.Tasks/Tests/Xamarin.ProjectTools/../../../../bin/TestRelease/XABuildPaths.cs' could not be found.

The fix is to update `make prepare-paths` so that it generates
`bin/Test$(Configuration)/XABuildPaths.cs` for *all*
`$(CONFIGURATIONS)` values.