[Xamarin.Android.Build.Test] Rework CreateFauxAndroidSdkDirectory to take ApiInfo[]. (#1566)
Context: https://github.com/xamarin/xamarin-android/issues/1511

For most tests we already define an `ApiInfo[]` to contain apis we
want to test against. This commit cleans up the helper methods to
make use of the same `ApiInfo[]` arrays to create the
`android-{Id}` directories.

It also adds unit tests to conver the fix for #1511 in efe81f56.