[linker] Update linker to be aware of jni marshal methods
Let linker preserve the jni marshal methods and also update the
registration methods we generate with `jnimarshalmethod-gen`.

We want the linker to preserve only jni marshal methods for methods
which are not linked away. Thus we also need to update the
`__<$>_jni_marshal_methods::__RegisterNativeMembers` method to only
register the remaining methods.

For example in our XA template application, the number of registered
methods for `Android.App.Activity` class goes down to 31 from 241.

This commit still keeps the `Get*Handler` methods in the linked
assembly. Once we have the marshal methods for all the code, we can
remove these easily in `DoAdditionalMethodProcessing` by returning
early, before `PreserveRegisteredMethod` is called or even remove that
call.