[xabuild.exe] [tests] use Microsoft.Build.Locator
In `Xamarin.ProjectTools` and `xabuild.exe`, we had some code looking
for Visual Studio 2017 installations on Windows.

Unfortunately we weren't doing it "the right way" and were merely
checking a few directories within `%ProgramFiles(x86)%`.

Making use of the `Microsoft.Build.Locator` NuGet package, allows for
simpler (and more correct) code for finding instances of Visual
Studio. I took the approach of choosing the "newest" installation,
which works well on my machine because it chooses my 15.7 preview
installation instead of the 15.6 one. Note that this NuGet only works
on Windows, so we can only use it in the cases we had the existing
Windows platform-specific code.

To enable a way to "choose" an instance of Visual Studio, I also added
checks for the `%VSINSTALLDIR%` environment variable. This allows
usage of the *Developer Command Prompt for VS 2017* or just setting
the environment variable system-wide.

Sadly the NuGet of `Microsoft.Build.Locator` has an incompatible
license:

http://go.microsoft.com/fwlink/?LinkId=329770

Instead, we are using the source found on Github that is MIT licensed:

https://github.com/Microsoft/MSBuildLocator/blob/master/LICENSE

Other changes:
- Had to create a "shim" project to link-in the source files for
  `Microsoft.Build.Locator`
- Had to raise `TargetFrameworkVersion` to at least 4.6, so I chose
  4.6.2 to match Xamarin.Android.Build.Tests
- The IDE made some other adjustments in `*.csproj` files, but they
  appeared to be "good" formatting changes or corrections.
- Updated `setup-windows.exe` to also take advantage of
  `Microsoft.Build.Locator`, so you can now use a plain elevated
  `cmd.exe`