[Xamarin.Android.Build.Tasks] coded warnings for <BuildApk /> task
Context: https://github.com/xamarin/xamarin-android/issues/1560

`<BuildApk />` is in need of two warning codes to be fully fleshed out:
- `XA4301` is now fully documented and tested
- `XA0107` is *new* and now fully documented. `<ResolveAssemblies />`
also needed to use this warning code. I was not able to write a test
that produces this warning, since it does not happen under normal
circumstances

Other changes:
- The `BuildWithNativeLibraries` test was using `StringAssertEx.ContainsText`
this calls `Assert.Pass` internally (which throws) so the subsequent
assertions were not being checked.
- Added missing overloads for `LogCodedError` and `LogCodedWarning` in
`AsyncTask`.