[Xamarin.Android.Build.Tasks] Stop `_LinkAssembliesNoShrink` running all the time.
This has been a long standing issue that we never caught...
The `Output` on the `_LinkAssembliesNoShrink` target was
incorrect. It was using `$(MonoAndroidIntermediateDir)`
rather than `$(MonoAndroidIntermediateAssemblyDir)`.

As a result it was always looking for the output files in the
project root directory.. They never existed so the target
ALWAYS ran!.