[Xamarin.Android.Build.Tasks] _UpdateAndroidResgen always runs. (#1602)
Commit a50e00bb broke the build in such that `_UpdateAndroidResgen`
always runs. This is because the resource file last modified date
was always newer than the `R.cs.flag`.

As a result of reading the fix to set the last write, this then broke
the `_GenerateAndroidResourceDir` target, because `build.props` ended
up being newer than the resource files in `@(_AndroidResourceDest)`,
causing the target to run when it shouldn't. So we also introduce an
`$(_AndroidResFlagFile)` to control when we need to generate the
resource directory in the `$(IntermediateOutputPath)`.

Unit tests have been updated to make sure these target do not run
when they are no supposed to.