[mono-sdks] Pass the Host{Cc,Cxx}{32,64} parameters (Take 2!) (#1611)
Will hopefully allow LLVM to be configured on Linux...

The problem with the *previous* attempt to fix LLVM on Linux --
9a985136 -- was that it broke the build (3992ee73), apparently
because `CC` and `CXX` were *always* overridden, including on the
MXE/WinAPI builds:

	make -j4 package-llvm-llvmwin32 CONFIGURATION=debug IGNORE_PROVISION_MXE=true IGNORE_PROVISION_ANDROID=true IGNORE_PACKAGE_LLVM=true ANDROID_TOOLCHAIN_DIR="/Users/builder/android-toolchain" ANDROID_TOOLCHAIN_CACHE_DIR="/Users/builder/android-archives" ANDROID_TOOLCHAIN_PREFIX="/Users/builder/android-toolchain/toolchains" ANDROID_BUILD_TOOLS_VERSION="28-rc1" ANDROID_BUILD_TOOLS_DIR="28.0.0-rc1" LLVM_SRC="…/xamarin-android/src/mono-runtimes/../../external/llvm" MXE_PREFIX_DIR="/Users/builder/android-toolchain" MXE_SRC="…/xamarin-android/src/mono-runtimes/../../external/mxe" CC="" CXX=""

The fix for this is to override `CC` and `CXX` *only* on *non-MXE*
builds. This in turn is done by using new
`%(_LlvmRuntime.PackageOptions)` metadata which contains the
appropriate `CC="$(HostCc32)" CXX="$(HostCxx32)"`/etc. values for
Linux, but is not set at all for MXE builds. This prevents `CC` and
`CXX` from being overriden during `make package-llvm-llvmwin32`,
allowing the build to complete on macOS.