[Xamarin.Android.Build.Tasks] Fix test broken by new Binding system
Commit ab3773cf added support for a new Binding system. It
broke on of our updated unit tests though. If a resource is
updated but not changed the resulting binding generator should
NOT produce an updated binding cs file. If it does the result
is the `CoreCompile` target ends up running and a whole host
of other targets then run as a result including `_LinkAssembliesNoShrink`.

What we should be doing is generating to a temp file and then
only update the binding if there has been a change.