[Xamarin.Android.Build.Tasks] Xamarin Android Lint checks fail on Windows
Fixes #1374

Turns out `lint.bat --version` on windows returns `unknown version`.
As a result none of the tests we want to be disabled are disabled.
So we can work around this by extracting the version number from
the filename of the .jar file that is being used.

	lint-26.0.0-dev.jar

So if we find `unknown` we will try to figure out the version
from the filename of the .jar. Failing that lets default to
at least 1.0 so that our base checks are disabled correctly.