[performance] initial MSBuild project timing (#1569)
"Performance", in various guises, is an ongoing focus. For example,
we track [application startup times][time-plots].

[time-plots]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/plot/Tests%20times/

The next area we want to concentrate on is build and deployment time,
and to do so we need to start collecting information so we know how
we're doing, and whether anything is getting worse.

The first step in tracking our build times is to track the build
times for some of our existing unit tests projects, and plot those
on Jenkins. This is done by using a custom MSBuild logger
`Xamarin.Android.Tools.BootstrapTasks.TimingLogger` which records
detailed timing information into an XML file:

	<builds>
	  <build
	      start-ticks="636607852701529939"
	      start-threadid="1"
	      id="HelloWorld-FreshBuild"
	      commit="8dcd59e"
	      description="A fresh build after a clean checkout for HelloWorld"
	      end-ticks="636607852899547509"
	      end-threadid="1"
	      elapsed="00:00:19.8017570"
	      succeeded="True">
	    <toolsets>
	      <toolset version="15.0" path="C:\Program Files (x86)\Microsoft Visual Studio\2017\Enterprise\MSBuild\15.0\Bin" />
	    </toolsets>
	    <files>
	      <file id="1" path="C:\Users\jopepper\Desktop\Git\xamarin-android\samples\HelloWorld\HelloWorld.csproj" />
	    </files>
	    <project
	        start-ticks="636607852707145402"
	        start-threadid="6"
	        file-id="1"
	        end-ticks="636607852899326841"
	        end-threadid="6"
	        elapsed="00:00:19.2181439"
	        succeeded="True">
	      <target
	          start-ticks="636607852707365985"
	          start-threadid="6"
	          name="RedirectMonoAndroidSdkPaths"
	          file-id="2"
	          end-ticks="636607852708932958"
	          end-threadid="6"
	          elapsed="00:00:00.1566973"
	          succeeded="True">
	        <task
	          start-ticks="636607852708755105"
	          start-threadid="6"
	          name="SetVsMonoAndroidRegistryKey"
	          file-id="2"
	          end-ticks="636607852708895478"
	          end-threadid="6"
	          elapsed="00:00:00.0140373"
	          succeeded="True"
	        />
	      </target>
	    </project>
	  </build>
	</builds>


These timing XML files from multiple builds are aggregated by the new
`<ProcessMSBuildTiming/>` task, which generates a CSV file containing
the `//*/@elapsed` attribute values for us by the
[Jenkins Plots plugin][jenkins-plots].

[jenkins-plots]: https://wiki.jenkins.io/display/JENKINS/Plot+Plugin

The following projects have their build times collected:

  - `samples/HelloWorld`
  - `tests/Xamarin.Forms-Performance-Integration`

Across four types of builds:

  - Fresh build (simulated fresh checkout)
  - Second build (no changes)
  - Touch C# file, build again
  - Touch Android resource XML file, build again

We are timing the `SignAndroidPackage` and `Install` targets for each
of these build types.

Results are aggregated into `TestResult-MSBuild-times.csv` for
processing on Jenkins. After this gets merged into master, there will
be some configuration needed to setup the new plot.

Example results, in the aggregated CSV file:

	| HelloWorld-FreshBuild-Debug | HelloWorld-FreshInstall-Debug | XF-FreshBuild-Debug | XF-FreshInstall-Debug |
	| ---:                        | ---:                          | ---:                | ---:                  |
	| 00:00:16.3538291            | 00:00:04.7940071              | 00:00:27.5263076    | 00:00:22.6720137      |

Also included in the table, will be the `TouchCSharp` and
`TouchAndroidResource` builds for both of the `HelloWorld` and `XF`
projects for both `SignAndroidPackage` and `Install` targets.

Other changes:

  - I let Visual Studio 2017 re-sort/fixup `Xamarin.Android.sln`
  - Split up `timing.csproj` into `timing.targets` and `timing.projitems`
  - Updated `Makefile` to include a `run-performance-tests` target.

`TimingLogger` is courtesy of @grendello.