[build] Avoid using `$(wildcard) on generated files
Commit 7f56c949 attempted to fix `make package-oss` on Linux, noting:

> In the absense of actually understanding what's wrong and fixing the
> real problem, instead print out some debug spew if the `cp` operation
> *would* fail -- because `$(firstword $(wildcard ...))` returns "" --
> and fallback to an attempted "hardcoded" `cp` operation.

We now have a plausible explanation which matches the facts on hand:
[`$(wildcard)` should be used with care][wildcard], because it's
behavior is dependant upon filesystem caching done by **make**.

[wildcard]: https://www.cmcrossroads.com/article/trouble-wildcard

> The unexpected, and apparently inconsistent, results above are
> because GNU Make contains its own cache of directory entries.
> `$(wildcard)` reads from that cache (and not directly from disk like
> `ls`) to get its results. When that cache is filled is vital to
> understanding the results the `$(wildcard)` will return.

In short, `$(wildcard)` should *not* be used with files that are
potentially generated during the "current" `make` invocation.
Use with "static" file lists, or files generated by *unrelated* `make`
invocations should be fine, but not within the *same*.

Review `$(wildcard)` use and use alternate mechanisms when we want to
use files which were potentially built within the same `make` process.