[Xamarin.Android.Build.Tasks] Fix XA5300 error reporting
When the Android SDK cannot be found, an XA5205 error *should* have
been reported.

Unfortunately for @directhex on Linux, that wasn't the case:

	$ msbuild Project.csproj
	...
	error MSB4018: The "ResolveSdks" task failed unexpectedly.
	error MSB4018: System.InvalidOperationException: Could not determine Android SDK location. Please provide `androidSdkPath`.
	error MSB4018:   at Xamarin.Android.Tools.AndroidSdkInfo..ctor (System.Action`2[T1,T2] logger, System.String androidSdkPath, System.String androidNdkPath, System.String javaSdkPath)
	error MSB4018:   at Xamarin.Android.Tasks.MonoAndroidHelper.RefreshAndroidSdk (System.String sdkPath, System.String ndkPath, System.String javaPath)
	error MSB4018:   at Xamarin.Android.Tasks.ResolveSdks.RunTask ()
	error MSB4018:   at Xamarin.Android.Tasks.ResolveSdks.Execute ()
	error MSB4018:   at Microsoft.Build.BackEnd.TaskExecutionHost.Microsoft.Build.BackEnd.ITaskExecutionHost.Execute ()
	error MSB4018:   at Microsoft.Build.BackEnd.TaskBuilder+<ExecuteInstantiatedTask>d__26.MoveNext ()

The problem is due to commit b16ee559, which integrated
xamarin-android-tools and introduced use of `AndroidSdkInfo`.

`AndroidSdkInfo` *requires* that it be provided or that it be able to
*find* an Android SDK or Java JDK. If neither can be found, then the
`AndroidSdkInfo` constructor will throw an
`InvalidOperationException`, precisely as seen above.

This semantic change was overlooked, and generally wouldn't be
observed unless building on a "clean machine" that didn't have a
previous Xamarin.Android installation.

Update `MonoAndroidHelper.RefreshAndroidSdk()` so that the
`InvalidOperationException` is appropriately converted to an MSBuild
error, improving the user experience.

Additionally, change the error code from XA5205 (originally
"missing `aapt`") to XA5300 ("Android SDK not found") and add
documentation for the error message.