[Xamarin.Android.Build.Tasks] Xamarin Android Lint checks fail on Windows (#1627)
Fixes: https://github.com/xamarin/xamarin-android/issues/1374

Context: https://github.com/xamarin/xamarin-android/issues/1560

`lint.bat --version` on Windows returns `unknown version`.  As a
result none of the tests we want to be disabled are disabled.
We can work around this by extracting the version number from the
filename of the `.jar` file that is being used:

	lint-26.0.0-dev.jar

If we find `unknown` we will try to figure out the version from the
filename of the `.jar`.  Failing that let's default to at least 1.0
so that our base checks are disabled correctly.

Additionally, add documentation for the XA0102, XA0103, and XA0108
build messages.